backend/requirements.txt
asgiref==3.3.4
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
numpy==1.21.0
pandas==1.2.5
psycopg2==2.9.1
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
